disgaea

+++ tools/bash/video-32mp4.sh
@@ -24,6 +24,7 @@ function setsize {
 
 af=''
 srt=''
+fmt='mp4'
 SECONDS=0
 ffprobe=(
 	ffprobe
@@ -41,36 +42,62 @@ while [ "$1" ]; do
 	shift
 
 	if [ -f "$t1" ]; then
-		mime=$(file  --brief  --mime-type  "$t1")
-		tmp="/tmp/$$.mp4"
-		case "$mime" in
-			'video/'* | 'audio/'*)
-				[ -f "$tmp" ] && rm -vf "$tmp"
-				setsize $(${ffprobe[@]}  "$t1")
+		mime=$(file  --brief  --mime-type  "$t1" | grep 'video/')
+		[ "$mime" ] || continue
 
-				echo "s=$size  af=$af  srt=$srt"
-				nice -n 19  ffmpeg -y \
-					-v 0            \
-					-i "$t1"        \
-					$srt            \
-					-s $size        \
-					-aspect $orin   \
-					-vcodec libx264 \
-					-q:v 0          \
-					-b:a 24k        \
-					-r 15 -g 150    \
-					-ac 1 -ar 44100 \
-					$af             \
-					-max_muxing_queue_size 2048 \
-					"$tmp"
-				[ -s "$tmp" ] || rm -v  "$tmp"  # remove 0 byte file
-				[ -f "$tmp" ] && mv -vf "$tmp"  "$bas-$size".mp4
-				;;
-		esac
+		if [[ $fmt == 'mp4' ]]; then
+			tmp="/tmp/$$.mp4"
+			[ -f "$tmp" ] && rm -vf "$tmp"
+
+			setsize $(${ffprobe[@]}  "$t1")
+			echo "[$#] [mp4] s=$size  af=$af  srt=$srt"
+
+			nice -n 19  ffmpeg -y \
+				-v 0            \
+				-i "$t1"        \
+				$srt            \
+				-s $size        \
+				-aspect $orin   \
+				-vcodec libx264 \
+				-q:v 0          \
+				-b:a 24k        \
+				-r 15 -g 150    \
+				-ac 1 -ar 44100 \
+				$af             \
+				-max_muxing_queue_size 2048 \
+				"$tmp"
+			[ -s "$tmp" ] || rm -v  "$tmp"  # remove 0 byte file
+			[ -f "$tmp" ] && mv -vf "$tmp"  "$bas-$size".mp4
+		fi
+		if [[ $fmt == 'png' ]]; then
+			tmp=$(tr -c '[0-9a-zA-Z]' \_ <<< "$t1")
+			[ -d "$tmp" ] && rm -vfr "$tmp"
+			mkdir -p "$tmp"
+
+			setsize $(${ffprobe[@]}  "$t1")
+			echo "[$#] [png] s=$size"
+
+			#     99 =          1:39
+			#    999 =         16:39
+			#   9999 =       2:46:39
+			#  99999 =   1d  3:46:39
+			# 999999 =  11d 13:46:39
+			nice -n 19  ffmpeg -y \
+				-v 0          \
+				-i "$t1"      \
+				-s $size      \
+				-aspect $orin \
+				-qscale 0     \
+				-r 1          \
+				-an           \
+				"$tmp"/%05d.png
+		fi
 	else
 		case "${t1:2}" in
 			'af' )   af='-af loudnorm=I=-14:TP=-1';;
 			'srt')  srt='-map 0:v  -map 0:a  -map 0:s? -scodec copy';;
+			'mp4')  fmt='mp4';;
+			'png')  fmt='png';;
 			*)
 				let s=${t1:2}*1
 				if (( $s < 1 )); then
